Refactor Social Links Component
