+++ src/components/header/NavBar.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
-import SocialLinksBar from '../SocialLinksBar';
+import SocialLinksBar from '../shared/SocialLinksBar';
 import ContactPopUpForm from '../quoteform/ContactPopUp';
 
 function NavBar() {
